requirements.txt
crewai
crewai-tools
python-dotenv
pyyaml
pydantic
pytest